test: relocate test setup into per-package files
Each package now owns its test infrastructure. The package-local suite
is the canonical entry point (`cd packages/<x> && bun test`); the root
`bunfig.toml` stays purely as a convenience composer.

Layout per package:
  packages/state/   bunfig.toml + test.setup.ts (matchers, mock helpers,
                    Context.root.pop afterEach - no DOM)
  packages/react/   bunfig.toml + test.setup.ts (registers happy-dom
                    inline at top, imports state's test.setup, requires
                    @testing-library/react via `require()` so it loads
                    after happy-dom has mutated globalThis)

The require() is deliberate: ESM hoists `import` statements above any
top-level code, so the `register()` call has to happen before any line
that pulls in @testing-library/dom transitively. Documented inline.

Root bunfig.toml now just preloads `./packages/react/test.setup.ts`,
which transitively brings in the state setup. Drop the previous root
`test.preload.ts` and `test.setup.ts`. Preact's `test.ts` is retargeted
at `../state/test.setup` (one-line change) so its existing inline
happy-dom setup keeps working unchanged.

All four entry points pass: root (657), packages/state (452),
packages/react (173), packages/preact (32).
