+++ packages/preact/test.ts
@@ -27,7 +27,7 @@ export const vi = {
   spyOn
 };
 
-export { mockError, mockPromise, mockWarn } from '../../test.setup';
+export { mockError, mockPromise, mockWarn } from '../state/test.setup';
 export {
   act,
   fireEvent,
