--- test.preload.ts
@@ -1,3 +1,0 @@
-import { GlobalRegistrator } from '@happy-dom/global-registrator';
-
-if (!GlobalRegistrator.isRegistered) GlobalRegistrator.register();
